--story=858321519 插件上架支持代码扫描和质量红线

+++ src/backend/ci/ext/tencent/store/biz-store-tencent/src/main/kotlin/com/tencent/devops/store/service/common/impl/TxStoreCodeccValidateServiceImpl.kt
@@ -241,7 +241,7 @@ class TxStoreCodeccValidateServiceImpl @Autowired constructor(
         } else {
             throw ErrorCodeException(errorCode = CommonMessageCode.ERROR_CLIENT_REST_ERROR)
         }
-        return String.format(scoreFormat, score).toDouble()
+        return String.format(scoreFormat, score.toDouble()).toDouble()
     }
 
     /**
